New version 1.5.0:
Removed the WPngImage_libpng.cc and WPngImage_lodpng.cc and moved their content into the WPngImage.cc file.
The choice of library is now done with a preprocessor macro (with the default being to use lodepng.)

Updated lodepng to the latest version.

Added a new utility utils/WVerticalBarGraph for drawing simple bar graphs. Moved the WColorSequence
utility into a subdirectory of utils.